feat(routing): move /home into (main) Route Group with shared layout
