+++ src/components/AppSidebar.tsx
@@ -18,9 +18,9 @@ import Link from 'next/link'
 import { usePathname, useRouter } from 'next/navigation'
 import { useCallback, useMemo, useState } from 'react'
 
-import { Category } from '@/app/home/_components/Category'
-import { CategoryManageDialog } from '@/app/home/_components/CategoryManageDialog'
-import { LogoutButton } from '@/app/home/_components/LogoutButton'
+import { Category } from '@/app/(main)/home/_components/Category'
+import { CategoryManageDialog } from '@/app/(main)/home/_components/CategoryManageDialog'
+import { LogoutButton } from '@/app/(main)/home/_components/LogoutButton'
 import { useIsElectron } from '@/components/auth/ElectronLoginForm'
 import { ThemeSelectorMenuItem } from '@/components/ThemeSelectorMenuItem'
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
